+++ examples/addons/drawing/draw_cad.py
@@ -11,7 +11,7 @@ import ezdxf
 from ezdxf import recover
 from ezdxf.addons.drawing import RenderContext, Frontend
 from ezdxf.addons.drawing.matplotlib import MatplotlibBackend
-from ezdxf.addons.drawing import fonts
+from ezdxf.tools import fonts
 
 # Setup fonts - this is not done automatically, because this may take a long
 # time and is not important for every user.