+++ tests/test_08_addons/test_811_drawing_frontend.py
@@ -5,7 +5,7 @@ import pytest
 import ezdxf
 from ezdxf.addons.drawing import Frontend, RenderContext, Properties
 from ezdxf.addons.drawing.backend import Backend
-from ezdxf.addons.drawing.text import FontMeasurements
+from ezdxf.tools.fonts import FontMeasurements
 from ezdxf.document import Drawing
 from ezdxf.entities import DXFGraphic
 from ezdxf.render.forms import cube